Refactoring
- Rename folders
- Refactor for Django 1.6

Some static paths/URLs in settings are still there. Futurice specific
graphics should be removed.
