+++ filesharing/filesharing/urls.py
@@ -1,0 +1,16 @@
+from django.conf.urls import patterns, include, url
+
+urlpatterns = patterns('',
+    # Root view
+    (r'^$', 'upload.views.index'),
+    # Upload url
+    (r'^(?P<folder>\d+)/$', 'upload.views.upload'),
+    #Zipping url
+    (r'zip/(?P<folder>\d+)/$', 'upload.views.zip'),
+    # Get zip
+    (r'^(?P<zip>\d+\.\w+)$', 'upload.views.getzip'),
+    #Send mail + sms
+    (r'send/$', 'upload.views.send'),
+    # Password prompt
+    (r'ask/(?P<requestedFilename>\d+\.\w+)$', 'upload.views.passwordCheck'),
+)
