+++ filesharing/templates/upload/base_ask.html
@@ -1,4 +1,4 @@
-{% extends "futuUpload/base.html" %}
+{% extends "upload/base.html" %}
 {% load static %}
 
 {% block title %}{{ block.super }}: Password{% endblock %}
